S15.f.1: extract plugin.ts redact + AMID-cache helpers (7139 → 6974 LOC)
First sub-slice of the S15.f plugin.ts decomposition train. Lifts the
shared utility primitives out of plugin.ts so subsequent slices can
move tools without dragging plugin.ts state with them.

cli/src/core/log-redact.ts (40 LOC):
  redactSecrets(m), sanitizeLog(s, maxLen)

cli/src/core/amid-cache.ts (213 LOC):
  Module state: amidToName, nameToAmid, nameToAmidTs,
                parentTrustedAmids, peerSigningKeys, AMID_CACHE_TTL_MS
  Helpers:      getCachedAmid, setCachedAmid, pickFreshestRegistryMatch
  Resolvers:    resolveAmidByName, resolveAmidToName, resolveSigningKey
                (take routerUrl as parameter to avoid circular dep)

plugin.ts re-exports redactSecrets to preserve the legacy import
surface. Thin no-arg wrappers in plugin.ts thread routerUrl into the
resolvers so existing call-sites stay byte-identical.

Behavioral change: NONE. Maps are exported as live Map/Set instances;
all 89 cross-references in plugin.ts continue to read/mutate the
same singletons.

Verification:
  - tsc --noEmit clean / lint 24 (+2 any-disables in amid-cache) / build clean
  - npm test: 454 pass / 2 skipped pre-existing (same as baseline)
  - Cross-cutting AMID maps verified by full plugin.test.ts pass — the
    suite exercises mesh handoff/spawn/send/inbox which all read/mutate
    the extracted maps end-to-end.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f1.md
